+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# next (XX.XX.2021)
+
+### Breaking Changes
+
+-   **terra-note-editor**
+
+    -   deprecated component is removed.
+
+-   **terra-code-editor**
+
+    -   deprecated component is removed.
+
+-   **dependencies**
+    -   packages `quill` `htmlhint` `ngx-quill` are removed.
+
 # 11.5.1 (30.06.2021)
 
 ### Bug fixes